% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choroplethrOsakaCity-package.R
\docType{package}
\name{choroplethrOsakaCity}
\alias{choroplethrOsakaCity}
\alias{choroplethrOsakaCity-package}
\title{A Map of Osaka City for choroplethr}
\description{
This package is for the use with the choroplethr package.
You can draw the choropleth maps of Osaka, the second largest city in Japan.
The original shp file is destributed by the City Planning Bureau of Osaka City,
under CC-BY 2.1 license. For more details, see \url{http://www.city.osaka.lg.jp/toshikeikaku/page/0000250227.html}
}
\seealso{
\url{http://www.city.osaka.lg.jp/toshikeikaku/page/0000250227.html}
}

