% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{osaka_city.code}
\alias{osaka_city.code}
\alias{to_osaka_ward_code}
\alias{to_osaka_ward_name}
\title{Convert ward code from/to ward name}
\usage{
to_osaka_ward_name(ward_code, to = c("kanji", "alphabet"))

to_osaka_ward_code(ward_name, from = c("kanji", "alphabet"))
}
\description{
Convert ward code from/to ward name
}

