% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OsakaCityChoropleth.R
\docType{data}
\name{osakaCityChoropleth}
\alias{osakaCityChoropleth}
\alias{osaka_city_chropleth}
\title{A Choropleth Class for Osaka City Map}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000a1a9590> 
 - attr(*, "name")= chr "osakaCityChoropleth_generator"
}}
\usage{
osakaCityChoropleth

osaka_city_chropleth(df, title = "", legend = "", num_colors = 7,
  zoom = NULL)
}
\description{
A Choropleth Class for Osaka City Map
}
\keyword{datasets}

